requirements.txt
pandas
jieba
pkuseg
numpy
python-dateutil
gensim